mv toolkit to tool_manager
